# src/backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: SpotiBye API
  description: Cloudflare Workers backend for Spotify playlist analysis and export functionality
  version: 1.0.0
  contact:
    name: SpotiBye Team
    email: support@spotibye.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://spotibye-api.workers.dev
    description: Production server
  - url: https://spotibye-api-dev.workers.dev
    description: Development server
  - url: http://localhost:8787
    description: Local development server
paths:
...
  /spotify/playlists:
    get:
      tags:
        - Spotify Data
      summary: Get user's playlists
      description: Retrieves all playlists for the authenticated user
      operationId: getUserPlaylists
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 50
          description: Maximum number of playlists to return
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
          description: Offset for pagination
      responses:
        '200':
          description: Playlists retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistsResponse'
              examples:
                success:
                  summary: Successful playlists retrieval
                  value:
                    data:
                      playlists:
                        - id: "playlist-123"
                          name: "My Favorite Songs"
                          tracks_count: 100
                          owner: "Test User"
                          images:
                            - url: "https://example.com/image.jpg"
                              height: 300
                              width: 300
                      total: 1
                      limit: 50
                      offset: 0
        '401':
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /spotify/playlists/{id}:
    get:
      tags:
        - Spotify Data
      summary: Get playlist details
      description: Retrieves detailed information about a specific playlist
      operationId: getPlaylist
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Spotify playlist ID
      responses:
        '200':
          description: Playlist details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistResponse'
              examples:
                success:
                  summary: Successful playlist retrieval
                  value:
                    data:
                      playlist:
                        id: "playlist-123"
                        name: "My Favorite Songs"
                        description: "My favorite tracks"
                        tracks_count: 100
                        owner: "Test User"
                        followers: 42
                        public: true
                        collaborative: false
                        images:
                          - url: "https://example.com/image.jpg"
                            height: 300
                            width: 300
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Playlist not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /spotify/playlists/{id}/tracks:
    get:
      tags:
        - Spotify Data
      summary: Get playlist tracks
      description: Retrieves all tracks from a specific playlist
      operationId: getPlaylistTracks
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Spotify playlist ID
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
          description: Maximum number of tracks to return
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
          description: Offset for pagination
      responses:
        '200':
          description: Tracks retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TracksResponse'
              examples:
                success:
                  summary: Successful tracks retrieval
                  value:
                    data:
                      tracks:
                        - id: "track-123"
                          name: "Example Song"
                          artists:
                            - name: "Example Artist"
                              id: "artist-123"
                          album:
                            name: "Example Album"
                            id: "album-123"
                          duration_ms: 180000
                          track_number: 1
                          explicit: false
                          preview_url: "https://example.com/preview.mp3"
                          external_urls:
                            spotify: "https://open.spotify.com/track/track-123"
                      total: 100
                      limit: 50
                      offset: 0
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Playlist not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /export/playlist/{id}:
    post:
      tags:
        - Export
      summary: Generate playlist export
      description: Creates an export file for the specified playlist
      operationId: exportPlaylist
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Spotify playlist ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportRequest'
            examples:
              csv:
                summary: CSV export request
                value:
                  format: "csv"
                  include_audio_features: true
                  include_analysis: false
              json:
                summary: JSON export request
                value:
                  format: "json"
                  include_audio_features: true
                  include_analysis: true
      responses:
        '200':
          description: Export generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportResponse'
              examples:
                success:
                  summary: Successful export generation
                  value:
                    data:
                      export_id: "export-123"
                      format: "csv"
                      estimated_size: 1024000
                      expires_at: "2023-12-13T23:00:00Z"
                      download_url: "/export/playlist/playlist-123/download?export_id=export-123"
        '400':
          description: Invalid export parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Playlist not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Export generation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /export/playlist/{id}/download:
    get:
      tags:
        - Export
      summary: Download export file
      description: Downloads a previously generated export file
      operationId: downloadExport
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Spotify playlist ID
        - name: export_id
          in: query
          required: true
          schema:
            type: string
          description: Export file ID
      responses:
        '200':
          description: Export file downloaded successfully
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
              examples:
                csv:
                  summary: CSV file download
                  value:
                    schema:
                      type: string
                      format: binary
                    headers:
                      Content-Disposition: "attachment; filename=\"playlist-export.csv\""
                      Content-Type: "text/csv"
        '400':
          description: Invalid export ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Export file not found or expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '410':
          description: Export file has expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from Spotify OAuth callback
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Error code
            message:
              type: string
              description: Human-readable error message
            details:
              type: object
              description: Additional error details
      required:
        - error
...
    PlaylistsResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            playlists:
              type: array
              items:
                $ref: '#/components/schemas/Playlist'
            total:
              type: integer
              description: Total number of playlists
            limit:
              type: integer
              description: Number of playlists returned
            offset:
              type: integer
              description: Offset for pagination
          required:
            - playlists
            - total
            - limit
            - offset
      required:
        - data
    Playlist:
      type: object
      properties:
        id:
          type: string
          description: Spotify playlist ID
        name:
          type: string
          description: Playlist name
        description:
          type: string
          description: Playlist description
        tracks_count:
          type: integer
          description: Number of tracks in playlist
        owner:
          type: string
          description: Playlist owner name
        followers:
          type: integer
          description: Number of followers
        public:
          type: boolean
          description: Whether playlist is public
        collaborative:
          type: boolean
          description: Whether playlist is collaborative
        images:
          type: array
          items:
            $ref: '#/components/schemas/Image'
      required:
        - id
        - name
        - tracks_count
        - owner
    Image:
      type: object
      properties:
        url:
          type: string
          format: uri
          description: Image URL
        height:
          type: integer
          description: Image height in pixels
        width:
          type: integer
          description: Image width in pixels
      required:
        - url
    PlaylistResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            playlist:
              $ref: '#/components/schemas/Playlist'
      required:
        - data
    TracksResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            tracks:
              type: array
              items:
                $ref: '#/components/schemas/Track'
            total:
              type: integer
              description: Total number of tracks
            limit:
              type: integer
              description: Number of tracks returned
            offset:
              type: integer
              description: Offset for pagination
          required:
            - tracks
            - total
            - limit
            - offset
      required:
        - data
    Track:
      type: object
      properties:
        id:
          type: string
          description: Spotify track ID
        name:
          type: string
          description: Track name
        artists:
          type: array
          items:
            $ref: '#/components/schemas/Artist'
        album:
          $ref: '#/components/schemas/Album'
        duration_ms:
          type: integer
          description: Track duration in milliseconds
        track_number:
          type: integer
          description: Track number on album
        explicit:
          type: boolean
          description: Whether track is explicit
        preview_url:
          type: string
          format: uri
          description: Preview URL for track
        external_urls:
          type: object
          properties:
            spotify:
              type: string
              format: uri
              description: Spotify URL for track
      required:
        - id
        - name
        - artists
        - album
        - duration_ms
    Artist:
      type: object
      properties:
        id:
          type: string
          description: Spotify artist ID
        name:
          type: string
          description: Artist name
      required:
        - id
        - name
    Album:
      type: object
      properties:
        id:
          type: string
          description: Spotify album ID
        name:
          type: string
          description: Album name
      required:
        - id
        - name
...
    ExportRequest:
      type: object
      properties:
        format:
          type: string
          enum: [csv, json, xlsx]
          description: Export format
        include_audio_features:
          type: boolean
          default: true
          description: Whether to include audio features
        include_analysis:
          type: boolean
          default: false
          description: Whether to include analysis data
      required:
        - format
    ExportResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            export_id:
              type: string
              description: Unique export file ID
            format:
              type: string
              description: Export format
            estimated_size:
              type: integer
              description: Estimated file size in bytes
            expires_at:
              type: string
              format: date-time
              description: When export file expires
            download_url:
              type: string
              description: URL to download the export
          required:
            - export_id
            - format
            - estimated_size
            - expires_at
            - download_url
      required:
        - data
...